Methods.py is its own thing now! File still funky
