+++ fit_continuous/__init__.py
@@ -2,20 +2,13 @@ from fastapi import Response
 from fastapi import Request
 import pandas as pd
 from tempfile import NamedTemporaryFile
-from fit_continuous.methods import table_upload
+#from fit_continuous.methods import table_upload
+from fit_continuous.methods import *
 import rpy2
 from rpy2 import robjects
 import json
 # use pandas to store CSV files
 import pandas as pd
-
-#from methods import testRInterface
-
-def testRInterface():
-    r = robjects.r
-    r('''
-    print('howdy from R')
-    ''')
 
 
 def store_the_table(table):
@@ -37,70 +30,6 @@ def store_the_tree(tree):
     # we know the R methods can read files, so lets write out the file
     with open("/tmp/tree_file.phy", mode = "w") as f:
         f.write(treeString)
-
-def run_method(params):
-    # pull the arguments from the JSON object
-    print(   'got to the run method'   )
-    column = params['column']
-    model = params['model']
-    stdError = params['stdError']
-    print('RUN METHOD: params were',column,model,stdError)
-    # now that we have all the data collected, run the R method
-    r = robjects.r
-    env = robjects.globalenv
-    env['tree_file'] = '/tmp/tree_file.phy'
-    env['table_file'] = '/tmp/table_file.csv'
-    env['column'] = column
-    env['model'] = model
-    env['stdError'] = stdError
-    env['modelfit_summary_file'] = '/tmp/modelfile.csv'
-    env['plot_file'] = '/tmp/plotfile.png'
-    r('''
-require(ape)
-require(geiger)
-require(phytools)
-print('R is ready to execute the fit continuous method')
-
-plotsize = 1000
-
-tree <- read.tree(tree_file)
-table <- read.csv(table_file, row.names = 1, check.names = FALSE)
-
-td <- treedata(tree, table)
-df <- as.data.frame(td$data)
-dat <- df[,column, drop = FALSE] # Note: originally called selectedColumn
-phy <- td$phy
-
-# stdError might come over as a character instead of a number
-stderror <- as.numeric(stdError)
-
-# If the user inputs a non-number, stdError will be NA
-# Just make SE = 0 in that case
-if(is.na(stderror)) {
-    stderror <- 0
-}
-
-result <- fitContinuous(phy = phy, dat = dat, SE = stderror, model = model)
-result <- t(as.data.frame(unlist(result$opt)))
-rownames(result) <- "Primary results"
-result <- cbind(result, stderror) # Just to double-check the SE
-
-# Can I just use write.csv as before?
-write.csv(result, modelfit_summary_file)
-
-# Before the plot is made, dat needs to be named numbers
-dat <- dat[,1]
-names(dat) <- rownames(table)
-
-# Can I use normal png saving here? I guess this will be replaced by vega anyway
-png(plot_file, width = plotsize, height = plotsize)
-phenogram(phy, dat, fsize = 0.8, color = "darkgreen")
-dev.off()
-
-    ''')
-    print('** need to collect result from R here')
-    # Does not recognize the modelfit_summary_file variable
-    return modelfit_summary_file, plot_file
 
 
 def init(app):
